refactor labeled points
